Move buildQuery to utils since it not related to the algorithms
