--- packages/elasticsearch-asset-apis/src/elasticsearch-reader-api/algorithms/utils.ts
@@ -1,2 +1,0 @@
-// TODO: add the common utils here
-export {};
